# api/api.yaml
openapi: 3.0.1
info:
  title: Cloudburst
  version: 0.0.0
servers:
  - url: /api/v1
paths:
  /targets:
    get:
      summary: List ScrapeTargets
      operationId: ListScrapeTargets
      tags:
        - targets
      responses:
        '200':
          description: Listing successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ScrapeTarget"
  /targets/{target}/instances:
    put:
      summary: Update Instances for a ScrapeTarget
      operationId: UpdateInstances
      tags:
        - instances
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Instances'
      parameters:
        - in: path
          name: target
          required: true
          schema:
            type: string
      responses:
        200:
          description: Status update successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Instances"
    get:
      summary: Get Instances for a ScrapeTarget
      operationId: GetInstances
      tags:
        - instances
      parameters:
        - in: path
          name: target
          required: true
          schema:
            type: string
      responses:
        200:
          description: Listing successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Instances"
components:
  schemas:
    ScrapeTarget:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        query:
          type: string
        instanceSpec:
          $ref: '#/components/schemas/InstanceSpec'
    InstanceSpec:
      type: object
      properties:
        container:
          $ref: '#/components/schemas/ContainerSpec'
    ContainerSpec:
      type: object
      properties:
        name:
          type: string
        image:
          type: string
    Instance:
      type: object
      properties:
        name:
          type: string
        target:
          type: string
        endpoint:
          type: string
        active:
          type: boolean
        status:
          $ref: '#/components/schemas/InstanceStatus'
    InstanceStatus:
      type: object
      properties:
        agent:
          type: string
        status:
          type: string
          enum: ['unknown', 'pending', 'progress', 'failure', 'running', 'terminated']
        started:
          type: string
          format: 'date-time'
    Instances:
      type: array
      items:
        $ref: '#/components/schemas/Instance'