# api/server/go/go/model_instance.go
/*
 * Cloudburst
 *
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * API version: 0.0.0
 * Generated by: OpenAPI Generator (https://openapi-generator.tech)
 */

package openapi

type Instance struct {

	Name string `json:"name,omitempty"`

	Target string `json:"target,omitempty"`

	Endpoint string `json:"endpoint,omitempty"`

	Active bool `json:"active,omitempty"`

	Status InstanceStatus `json:"status,omitempty"`
}