# api/server/go/go/model_scrape_target.go
/*
 * Cloudburst
 *
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * API version: 0.0.0
 * Generated by: OpenAPI Generator (https://openapi-generator.tech)
 */

package openapi

type ScrapeTarget struct {

	Name string `json:"name,omitempty"`

	Description string `json:"description,omitempty"`

	Query string `json:"query,omitempty"`

	InstanceSpec InstanceSpec `json:"instanceSpec,omitempty"`
}